requirement.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
black==24.10.0
click==8.1.7
coverage==7.6.4
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.1.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.28.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
wcwidth==0.2.13
